Move builtins_*.go to cmd/containerd/builtins to make the files reusable
