+++ .golangci.yml
@@ -23,6 +23,11 @@ issues:
   max-issues-per-linter: 0
   max-same-issues: 0
 
+  # Only using / doesn't work due to https://github.com/golangci/golangci-lint/issues/1398.
+  exclude-rules:
+    - path: 'cmd[\\/]containerd[\\/]builtins[\\/]'
+      text: "blank-imports:"
+
 linters-settings:
   gosec:
     # The following issues surfaced when `gosec` linter
