+++ cmd/containerd/builtins/builtins.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package builtins
 
 // register containerd builtins here
 import (
