+++ cmd/containerd/builtins/builtins_linux.go
@@ -1,6 +1,3 @@
-//go:build !no_devmapper
-// +build !no_devmapper
-
 /*
    Copyright The containerd Authors.
 
@@ -17,6 +14,13 @@
    limitations under the License.
 */
 
-package main
+package builtins
 
-import _ "github.com/containerd/containerd/snapshots/devmapper/plugin"
+import (
+	_ "github.com/containerd/containerd/metrics/cgroups"
+	_ "github.com/containerd/containerd/metrics/cgroups/v2"
+	_ "github.com/containerd/containerd/runtime/v1/linux"
+	_ "github.com/containerd/containerd/runtime/v2/runc/options"
+	_ "github.com/containerd/containerd/snapshots/native/plugin"
+	_ "github.com/containerd/containerd/snapshots/overlay/plugin"
+)
