+++ cmd/containerd/builtins/builtins_windows.go
@@ -14,6 +14,11 @@
    limitations under the License.
 */
 
-package main
+package builtins
 
-import _ "github.com/containerd/zfs/plugin"
+import (
+	_ "github.com/containerd/containerd/diff/lcow"
+	_ "github.com/containerd/containerd/diff/windows"
+	_ "github.com/containerd/containerd/snapshots/lcow"
+	_ "github.com/containerd/containerd/snapshots/windows"
+)
