+++ cmd/containerd/builtins/zfs_linux.go
@@ -1,5 +1,5 @@
-//go:build darwin || freebsd || solaris
-// +build darwin freebsd solaris
+//go:build !no_zfs
+// +build !no_zfs
 
 /*
    Copyright The containerd Authors.
@@ -17,8 +17,6 @@
    limitations under the License.
 */
 
-package main
+package builtins
 
-import (
-	_ "github.com/containerd/containerd/snapshots/native/plugin"
-)
+import _ "github.com/containerd/zfs/plugin"
