+++ cmd/containerd/main.go
@@ -22,6 +22,9 @@ import (
 
 	"github.com/containerd/containerd/cmd/containerd/command"
 	"github.com/containerd/containerd/pkg/seed"
+
+	_ "github.com/containerd/containerd/cmd/containerd/builtins"
+	_ "github.com/containerd/containerd/pkg/cri"
 )
 
 func init() {
